Fix network loggers with MonitorState
